Rename to AspNetCoreDiagnosticSubscriber
AspNetCorePageLoadDiagnosticSubscriber -> AspNetCoreDiagnosticSubscriber
