+++ src/Elastic.Apm.AspNetCore/Elastic.Apm.AspNetCore.csproj
@@ -23,9 +23,6 @@
     <PackageReference Include="Microsoft.Extensions.Hosting.Abstractions" Version="3.0.0" />
   </ItemGroup>
   <ItemGroup>
-    <Folder Include="DiagnosticListener\" />
-  </ItemGroup>
-  <ItemGroup>
     <ProjectReference Include="..\Elastic.Apm.Extensions.Hosting\Elastic.Apm.Extensions.Hosting.csproj" />
     <ProjectReference Include="..\Elastic.Apm\Elastic.Apm.csproj" />
   </ItemGroup>
